- Implemented the Like module (like/unlike) - Still doing the dislike module - The changing image speed is so slow, later fix

+++ backend/route/dislikeRoute.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const Dislike = require("../model/like");
+const Dislike = require("../model/dislike");
 const rateLimit = require("express-rate-limit");
 const router = express.Router();
 
@@ -13,44 +13,61 @@ const dislikeLimiter = rateLimit({
 // Dislike
 router.post("/", dislikeLimiter, async (req, res) => {
   try {
-    const dislike = new Dislike({
-      postId: req.body.postId,
-      userId: req.body.userId,
-    });
+    const { postId, userId } = req.body;
 
-    const savedDislike = await dislike.save();
-    res.status(200).json(savedDislike);
+    /*
+    const existingLike = await Like.findOne({ postId, userId });
+    if (existingLike) {
+      return res.status(400).json({ message: "You have already liked this post." });
+    }*/
+
+    const newdislike = new Dislike({ postId, userId });
+    await newdislike.save();
+
+    const dislikes = await Dislike.find({ postId });
+    const dislikeNum = Dislikes.length;
+    const userIds = dislikes.map((like) => like.userId);
+
+    res.status(200).json({
+      postId,
+      dislikeNum,
+      userIds,
+    });
   } catch (err) {
-    res.status(401).json(err);
+    res.status(500).json({ message: "Debug 2" });
   }
 });
 
 // Undislike
 router.delete("/", dislikeLimiter, async (req, res) => {
-  await Dislike.deleteOne({
-    postId: req.body.postId,
-    userId: req.body.userId,
-  })
-    .then(() => {
-      res.json("deleted successfully");
-    })
-    .catch((err) => {
-      res.status(401).json(err);
-    });
+  const { postId, userId } = req.query;
+
+  await Dislike.deleteOne({ postId, userId })
+      .then(() => {
+        res.json("deleted successfully");
+      })
+      .catch((err) => {
+        res.status(401).json(err);
+      });
 });
 
 // Count dislike by postId
 router.get("/", async (req, res) => {
   try {
-    const dislikeNum = await Dislike.count({
-      postId: req.body.postId,
-    });
+    const { postId, userId } = req.query;
+
+    const dislikeNum = await Like.countDocuments({ postId });
+
+    const isDisliked = await Like.exists({ postId, userId });
 
     res.status(200).json({
-      dislikeNum: dislikeNum,
+      dislikeNum,
+      isDisliked,
+      postId,
+      userId,
     });
   } catch (err) {
-    res.status(401).json(err);
+    res.status(500).json({ message: "Debug 1" });
   }
 });
